Add owner list screen; Move classes related to derived owner selection to selection package
