--- app/src/main/java/io/gnosis/safe/ui/settings/owner/OwnerListFragment.kt
@@ -1,4 +1,0 @@
-package io.gnosis.safe.ui.settings.owner
-
-class OwnerListFragment {
-}
