+++ app/src/main/java/io/gnosis/safe/ui/settings/owner/list/OwnerListFragment.kt
@@ -1,0 +1,30 @@
+package io.gnosis.safe.ui.settings.owner.list
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import io.gnosis.safe.ScreenId
+import io.gnosis.safe.databinding.FragmentOwnerListBinding
+import io.gnosis.safe.di.components.ViewComponent
+import io.gnosis.safe.ui.base.fragment.BaseViewBindingFragment
+import javax.inject.Inject
+
+class OwnerListFragment : BaseViewBindingFragment<FragmentOwnerListBinding>() {
+
+    @Inject
+    lateinit var viewModel: OwnerListViewModel
+
+    override fun screenId() = ScreenId.OWNER_LIST
+
+    override fun inject(component: ViewComponent) {
+        component.inject(this)
+    }
+
+    override fun inflateBinding(inflater: LayoutInflater, container: ViewGroup?): FragmentOwnerListBinding =
+        FragmentOwnerListBinding.inflate(inflater, container, false)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
