+++ app/src/main/java/io/gnosis/safe/ui/settings/owner/selection/DerivedOwnerPagingSource.kt
@@ -1,4 +1,4 @@
-package io.gnosis.safe.ui.settings.owner.list
+package io.gnosis.safe.ui.settings.owner.selection
 
 import androidx.paging.PagingSource
 import io.gnosis.safe.utils.MnemonicAddressDerivator
@@ -6,7 +6,7 @@ import pm.gnosis.model.Solidity
 import timber.log.Timber
 
 
-class OwnerPagingSource(
+class DerivedOwnerPagingSource(
     private val derivator: MnemonicAddressDerivator,
     private val maxPages: Int
 ) : PagingSource<Long, Solidity.Address>() {
